Add a new 0.008% fee option (#52)
* Mutiply the fee options by 10 times

* Add tests for changing BPS and add a new feeOption

* Add FEE_UNITS

* Update _mintFee func

+++ contracts/periphery/KSRouter02.sol
@@ -181,7 +181,7 @@ contract KSRouter02 is IKSRouter02 {
                 tokenA,
                 tokenB,
                 ampBpsAndFeeBps[0],
-                uint16(ampBpsAndFeeBps[1])
+                uint24(ampBpsAndFeeBps[1])
             );
         }
         // if we add liquidity to an existing pool, this is an unamplifed pool
@@ -228,7 +228,7 @@ contract KSRouter02 is IKSRouter02 {
                 token,
                 weth,
                 ampBpsAndFeeBps[0],
-                uint16(ampBpsAndFeeBps[1])
+                uint24(ampBpsAndFeeBps[1])
             );
         }
         // if we add liquidity to an existing pool, this is an unamplifed pool
